Fix hold-forward macro bug
Had a weird issue that if you were holding S (backwards), then W (forwards) would not work. This went away when disabling the mod. This changes how holding-forward holds forward by mixin into the KeyboardInput class and pretending the key is pressed if the macro is enabled, rather than messing around with the keys themselves.