Simplify hideY since it's always true
hideY was previously set by the presence of a "civmc" resource file which was only included in civmc builds. However, hideY was changed to be hard-coded to true with 1c2b06f8f24d2d9dfa912977adb8e6ce9430617c. Since the previous commit removes civmc builds, the next step would be to simplify the code where the hideY-path is always taken.